unnecessary files removal + grouping the rest
