+++ lib/components/auth_screens/text_field.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 
-class MyTextField extends StatelessWidget {
+class AuthTextField extends StatelessWidget {
   final TextEditingController controller;
   final String hintText;
   final bool obscureText;
 
-  const MyTextField({
+  const AuthTextField({
     super.key,
     required this.controller,
     required this.hintText,
