+++ lib/screens/auth_screens/auth_screen.dart
@@ -1,4 +1,4 @@
-import 'package:app/screens/login_or_register_screen.dart';
+import 'package:app/screens/auth_screens/login_or_register_screen.dart';
 import 'package:app/screens/main_screen.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
