+++ lib/screens/auth_screens/login_or_register_screen.dart
@@ -1,5 +1,5 @@
-import 'package:app/screens/login_screen.dart';
-import 'package:app/screens/register_screen.dart';
+import 'package:app/screens/auth_screens/login_screen.dart';
+import 'package:app/screens/auth_screens/register_screen.dart';
 import 'package:flutter/material.dart';
 
 class LoginOrRegisterScreen extends StatefulWidget {
