+++ lib/screens/main_screen.dart
@@ -1,4 +1,4 @@
-import 'package:app/components/constants.dart' as constants;
+import 'package:app/utils/constants.dart' as constants;
 import 'package:app/screens/navbar_screens/events_screen.dart';
 import 'package:app/screens/navbar_screens/leagues_screen.dart';
 import 'package:app/screens/navbar_screens/profile_screen.dart';
