+++ lib/screens/navbar_screens/events_screen.dart
@@ -1,4 +1,3 @@
-import 'package:app/components/top_bar.dart';
 import 'package:flutter/material.dart';
 
 class EventsScreen extends StatefulWidget {
@@ -11,19 +10,27 @@ class EventsScreen extends StatefulWidget {
 class _EventsScreenState extends State<EventsScreen> {
   @override
   Widget build(BuildContext context) {
-    return SafeArea(
-      child: Column(
-        children: [
-          topBar(
-            context,
-            'lib/assets/images/Settings.svg',
-            () {
-              Navigator.pushNamed(context, '/settings');
-            },
+    return const Column(
+      children: [
+        // topBar(context,
+        //  'lib/assets/images/Settings.svg',
+        //  () {
+        //   //TODO navigate to settings screen
+        //  },
+        // ),
+        Expanded(
+          child: Center(
+            child: Text(
+              'Events go here',
+              style: TextStyle(
+                fontSize: 30,
+                fontWeight: FontWeight.bold,
+                color: Colors.black,
+              ),
+            ),
           ),
-          // const BetSelector(),
-        ],
-      ),
+        ),
+      ],
     );
   }
 }
