+++ lib/screens/navbar_screens/leagues_screen.dart
@@ -1,4 +1,3 @@
-import 'package:app/components/top_bar.dart';
 import 'package:flutter/material.dart';
 
 class LeaguesScreen extends StatefulWidget {
@@ -11,15 +10,15 @@ class LeaguesScreen extends StatefulWidget {
 class _LeaguesScreenState extends State<LeaguesScreen> {
    @override
   Widget build(BuildContext context) {
-    return Column(
+    return const Column(
       children: [
-        topBar(context,
-         'lib/assets/images/Settings.svg',
-         () {
-          //TODO navigate to settings screen
-         },
-        ),
-        const Expanded(
+        // topBar(context,
+        //  'lib/assets/images/Settings.svg',
+        //  () {
+        //   //TODO navigate to settings screen
+        //  },
+        // ),
+        Expanded(
           child: Center(
               child: Text(
                 'Leagues go here',
