+++ lib/screens/navbar_screens/profile_screen.dart
@@ -1,8 +1,8 @@
 import 'dart:io';
 
 import 'package:app/assets/translations.dart';
-import 'package:app/components/constants.dart' as constants;
-import 'package:app/components/nunito_text.dart';
+import 'package:app/utils/constants.dart' as constants;
+import 'package:app/components/other/nunito_text.dart';
 import 'package:app/components/profile_screen/gesture_detector_button.dart';
 import 'package:app/components/profile_screen/profile_pic.dart';
 import 'package:app/components/profile_screen/text_input_dialog.dart';
